New ways of collecting ontology files for hygiene tests and quick versions of FIBO.

+++ publisher/product/ontology/build.sh
@@ -95,8 +95,9 @@ function runHygieneTests() {
   #
   logRule "consistency-check: run for ..."
 
-  export DEV_SPEC="${DEV_SPEC:-About${ONTPUB_FAMILY^^}Dev.rdf}"
-  if [ -s "${source_family_root}/${DEV_SPEC}" ] ; then
+  # For now consistency check are turned off.  
+
+  if false && [ -s "${source_family_root}/${DEV_SPEC}" ] ; then
     rm -f ${TMPDIR}/{console.txt,ret.txt}
     logItem " DEV:${DEV_SPEC}" "$(getOntologyIRI < "${source_family_root}/${DEV_SPEC}")"
     if ${ONTOVIEWER_TOOLKIT_JAVA} --data "${source_family_root}/${DEV_SPEC}" \
@@ -109,8 +110,8 @@ function runHygieneTests() {
     fi
   fi
 
-  export PROD_SPEC="${PROD_SPEC:-About${ONTPUB_FAMILY^^}Prod.rdf}"
-  if [ -s "${source_family_root}/${PROD_SPEC}" ] ; then
+  
+  if false && [ -s "${source_family_root}/${PROD_SPEC}" ] ; then
     rm -f ${TMPDIR}/{console.txt,ret.txt}
     logItem "PROD:${PROD_SPEC}" "$(getOntologyIRI < "${source_family_root}/${PROD_SPEC}")"
     if ${ONTOVIEWER_TOOLKIT_JAVA} --data "${source_family_root}/${PROD_SPEC}" \
@@ -130,20 +131,29 @@ function runHygieneTests() {
   #
   # Get ontologies for Dev
   #
-  log "Merging all dev ontologies into one RDF file: $(logFileName ${tag_root}/DEV.ttl)"
-  "${JENA_ARQ}" $(find "${source_family_root}" -name "*.rdf" | grep -v "/etc/" | sed "s/^/--data=/") \
-    --query=/publisher/lib/echo.sparql \
-    --results=TTL > ${tag_root}/DEV.ttl
+  log "Merging all dev ontologies into one RDF file"
+  ${PYTHON3} ${SCRIPT_DIR}/lib/ontology_collector.py \
+     --root "${source_family_root}" \
+     --input_ontology "${source_family_root}/${DEV_SPEC}" \
+     --ontology-mapping "${source_family_root}/catalog-v001.xml" \
+     --output_ontology "${TMPDIR}/DEV.ttl"
 
+  success=$?
+  if [ "${success}" != 0 ] ; then return -1 ; fi
+  
   #
   # Get ontologies for Prod
   #
-  log "Merging all prod ontologies into one RDF file: : $(logFileName ${tag_root}/PROD.ttl)"
-  "${JENA_ARQ}" \
-    $(grep -r 'utl-av[:;.]Release' "${source_family_root}" | sed 's/:.*$//;s/^/--data=/' | grep -F ".rdf") \
-    --query=/publisher/lib/echo.sparql \
-    --results=TTL > ${tag_root}/PROD.ttl
+  log "Merging all prod ontologies into one RDF file"
+  ${PYTHON3} ${SCRIPT_DIR}/lib/ontology_collector.py \
+     --root "${source_family_root}" \
+     --input_ontology "${source_family_root}/${PROD_SPEC}" \
+     --ontology-mapping "${source_family_root}/catalog-v001.xml" \
+     --output_ontology "${TMPDIR}/PROD.ttl"
 
+  success=$?
+  if [ "${success}" != 0 ] ; then return -1 ; fi
+    
   logRule "Will run the following tests:"
 
   while read -r hygieneTestSparqlFile ; do
@@ -159,7 +169,7 @@ function runHygieneTests() {
     banner=$(getBannerFromSparqlTestFile "${hygieneTestSparqlFile}")
     logItem "Running test" "${banner}"
     ${JENA_ARQ} \
-      --data=${tag_root}/DEV.ttl \
+      --data=${TMPDIR}/DEV.ttl \
       --results=csv \
       --query="${hygieneTestSparqlFile}" | \
       sed 's/^\W*PRODERROR:/WARN:/g' | \
@@ -175,7 +185,7 @@ function runHygieneTests() {
     banner=$(getBannerFromSparqlTestFile "${hygieneTestSparqlFile}")
     logItem "Running test" "${banner}"
     ${JENA_ARQ} \
-      --data=${tag_root}/PROD.ttl \
+      --data=${TMPDIR}/PROD.ttl \
       --results=csv \
       --query="${hygieneTestSparqlFile}" | \
       sed 's/^\W*PRODERROR:/ERROR:/g' | \
@@ -681,16 +691,6 @@ __HERE__
   setProduct ontology || return $?
 
 
-  log "Merging all LCC ontologies into one ontology"
-  "${JENA_ARQ}" $(find "${INPUT}/LCC" -name "*.rdf" | sed "s/^/--data=/") \
-    --query=/publisher/lib/noimport_noontology.sparql \
-    --results=TTL > "${TMPDIR}/lcc.ttl"
-
-  log "Merging all external ontologies into one ontology"
-  "${JENA_ARQ}" $(find "${SCRIPT_DIR}/lib/ontologies" -name "*.rdf" | sed "s/^/--data=/") \
-    --query=/publisher/lib/noimport_noontology.sparql \
-    --results=TTL > "${TMPDIR}/external.ttl"
-
   log "Getting metadata"
   "${JENA_ARQ}" $(find "${source_family_root}" -name "MetadataFIBO.rdf" | grep -v "/etc/" | sed "s/^/--data=/") \
     --query=/publisher/lib/metadata.sparql \
@@ -699,32 +699,25 @@ __HERE__
   #
   # Get ontologies for Dev
   #
-  log "Merging all dev ontologies into one RDF file: $(logFileName ${tag_root}/dev.fibo-quickstart.ttl)"
-  "${JENA_ARQ}" $(find "${source_family_root}" -name "*.rdf" | grep -v "/etc/" | sed "s/^/--data=/") \
-    --query=/publisher/lib/noimport_noontology.sparql \
-    --results=TTL > ${TMPDIR}/pre_dev.fibo-quickstart.ttl
-
+  log "Merging all dev ontologies into one RDF file"
+  robot merge --input "${source_family_root}/${DEV_SPEC}" --output ${TMPDIR}/pre_dev.fibo-quickstart.owl
+ 
   ${JENA_ARQ} \
-    --data ${TMPDIR}/lcc.ttl \
     --data ${TMPDIR}/metadata.ttl \
-    --data ${TMPDIR}/pre_dev.fibo-quickstart.ttl \
+    --data ${TMPDIR}/pre_dev.fibo-quickstart.owl \
     --query=/publisher/lib/echo.sparql \
     --results=TTL > ${tag_root}/dev.fibo-quickstart.ttl
-
 
   #
   # Get ontologies for Prod
   #
-  log "Merging all prod ontologies into one RDF file: : $(logFileName ${tag_root}/prod.fibo-quickstart.ttl)"
-  "${JENA_ARQ}" \
-    $(grep -r 'utl-av[:;.]Release' "${source_family_root}" | sed 's/:.*$//;s/^/--data=/' | grep -F ".rdf") \
-    --query=/publisher/lib/noimport_noontology.sparql \
-    --results=TTL > ${TMPDIR}/pre_prod.fibo-quickstart.ttl
+  log "Merging all prod ontologies into one RDF file"
+  robot merge --input "${source_family_root}/${PROD_SPEC}" --output ${TMPDIR}/pre_prod.fibo-quickstart.owl
 
+  
   ${JENA_ARQ} \
-    --data ${TMPDIR}/lcc.ttl \
     --data ${TMPDIR}/metadata.ttl \
-    --data ${TMPDIR}/pre_prod.fibo-quickstart.ttl \
+    --data ${TMPDIR}/pre_prod.fibo-quickstart.owl \
     --query=/publisher/lib/echo.sparql \
     --results=TTL > ${tag_root}/prod.fibo-quickstart.ttl
